refactor: DI 모듈 패키지 구조 개선 및 재배치
- Firebase, JsonSerialization 모듈을 core 패키지로 이동
- NetworkQualifiers를 qualifier 패키지로 이동 및 관련 임포트 수정
