+++ app/src/main/java/com/boostcamp/and03/data/di/network/AladinBookLookUpNetworkModule.kt
@@ -2,8 +2,8 @@ package com.boostcamp.and03.data.di.network
 
 import com.boostcamp.and03.BuildConfig
 import com.boostcamp.and03.data.api.AladinBookLookUpApiService
-import com.boostcamp.and03.data.di.AladinBookLookUpOkHttpClient
-import com.boostcamp.and03.data.di.AladinBookLookUpRetrofit
+import com.boostcamp.and03.data.di.qualifier.AladinBookLookUpOkHttpClient
+import com.boostcamp.and03.data.di.qualifier.AladinBookLookUpRetrofit
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import dagger.Module
 import dagger.Provides
