+++ app/src/main/java/com/boostcamp/and03/data/di/qualifier/NetworkQualifiers.kt
@@ -1,4 +1,4 @@
-package com.boostcamp.and03.data.di
+package com.boostcamp.and03.data.di.qualifier
 
 import javax.inject.Qualifier
